Memory: move memory related files to subdir.
Only moves to be well detected by git
