+++ qtmips_cli/qtmips_cli.pro
@@ -27,6 +27,7 @@ unix: LIBS += \
         # --enable-new-dtags \
 
 INCLUDEPATH += $$PWD/..
+INCLUDEPATH += $$PWD/../qtmips_machine $$PWD/../qtmips_asm
 DEPENDPATH += $$PWD/../qtmips_machine $$PWD/../qtmips_asm
 QMAKE_CXXFLAGS += -std=c++14
 QMAKE_CXXFLAGS_DEBUG += -ggdb
