+++ qtmips_machine/instruction.cpp
@@ -35,7 +35,7 @@
 
 #include "instruction.h"
 
-#include "memory.h"
+#include "memory/backend/memory.h"
 #include "qtmipsexception.h"
 #include "utils.h"
