+++ qtmips_machine/memory/backend/peripheral.h
